+++ kiota_abstractions/request_information.py
@@ -9,6 +9,7 @@ from opentelemetry import trace
 from stduritemplate import StdUriTemplate
 
 from ._version import VERSION
+from .headers_collection import HeadersCollection
 from .method import Method
 from .request_option import RequestOption
 from .serialization import Parsable, SerializationWriter
@@ -54,7 +55,7 @@ class RequestInformation:
         self.query_parameters: Dict[str, QueryParams] = {}
 
         # The Request Headers
-        self.headers: Dict[str, Set[str]] = {}  # Use set to remove duplicates
+        self.headers: HeadersCollection = HeadersCollection()
 
         # The Request Body
         self.content: Optional[BytesIO] = None
@@ -95,46 +96,9 @@ class RequestInformation:
     @property
     def request_headers(self) -> Optional[Dict]:
         final = {}
-        for key, value in self.headers.items():
-            final[key] = ", ".join(value)
+        for key, values in self.headers.get_all().items():
+            final[key] = ', '.join(values)
         return final
-
-    def add_request_headers(
-        self, headers_to_add: Optional[Dict[str, Union[str, List[str]]]]
-    ) -> None:
-        """Adds headers to the request"""
-        if headers_to_add:
-            for key, value in headers_to_add.items():
-                lowercase_key = key.lower()
-                if lowercase_key in self.headers:
-                    if isinstance(value, list):
-                        self.headers[lowercase_key] = self.headers[lowercase_key].union(set(value))
-                    else:
-                        self.headers[lowercase_key].add(str(value))
-                else:
-                    if isinstance(value, list):
-                        self.headers[lowercase_key] = set(value)
-                    else:
-                        self.headers[lowercase_key] = {str(value)}
-
-    def try_add_request_header(self, key: str, value: str) -> bool:
-        """Try to add an header to the request if it's not already set"""
-        if key and value:
-            lowercase_key = key.lower()
-            if lowercase_key in self.headers:
-                return False
-            self.headers[lowercase_key] = {str(value)}
-            return True
-        return False
-
-    def remove_request_headers(self, key: str) -> None:
-        """Removes a request header from the current request
-
-        Args:
-            key (str): The key of the header to remove
-        """
-        if key and key.lower() in self.headers:
-            del self.headers[key.lower()]
 
     @property
     def request_options(self) -> Dict[str, RequestOption]:
@@ -227,13 +191,15 @@ class RequestInformation:
                 writer_func(None, values)
             self._set_content_and_content_type_header(writer, content_type)
 
-    def set_stream_content(self, value: BytesIO) -> None:
+    def set_stream_content(self, value: BytesIO, content_type: Optional[str] = None) -> None:
         """Sets the request body to be a binary stream.
 
         Args:
             value (BytesIO): the binary stream
         """
-        self.try_add_request_header(self.CONTENT_TYPE_HEADER, self.BINARY_CONTENT_TYPE)
+        if not content_type:
+            content_type = self.BINARY_CONTENT_TYPE
+        self.headers.try_add(self.CONTENT_TYPE_HEADER, content_type)
         self.content = value
 
     def set_query_string_parameters_from_raw_object(self, q: Optional[QueryParams]) -> None:
@@ -283,7 +249,7 @@ class RequestInformation:
         self, writer: SerializationWriter, content_type: Optional[str]
     ):
         if content_type:
-            self.try_add_request_header(self.CONTENT_TYPE_HEADER, content_type)
+            self.headers.try_add(self.CONTENT_TYPE_HEADER, content_type)
         self.content = writer.get_serialized_content()
 
     def _decode_uri_string(self, uri: Optional[str]) -> str: