+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2023-10-10
+
+### Added
+- Added a content type parameter to the set stream content method in request information.
+
+### Changed
+- Added dedicated HeadersCollection class to manage request headers.
+
 ## [0.8.7] - 2023-10-05
 
 ### Added